Cleanup interfaces and rename DynamoDbGateway
