+++ ContactDetailsApi.Tests/V1/UseCase/CreateContactUseCaseTests.cs
@@ -4,7 +4,8 @@ using ContactDetailsApi.V1.Boundary.Response;
 using ContactDetailsApi.V1.Domain;
 using ContactDetailsApi.V1.Domain.Sns;
 using ContactDetailsApi.V1.Factories;
-using ContactDetailsApi.V1.Gateways;
+using ContactDetailsApi.V1.Factories.Interfaces;
+using ContactDetailsApi.V1.Gateways.Interfaces;
 using ContactDetailsApi.V1.Infrastructure;
 using ContactDetailsApi.V1.UseCase;
 using FluentAssertions;
