+++ ContactDetailsApi.Tests/V1/UseCase/GetContactDetailsByTargetIdUseCaseTests.cs
@@ -3,7 +3,7 @@ using ContactDetailsApi.V1.Boundary.Request;
 using ContactDetailsApi.V1.Boundary.Response;
 using ContactDetailsApi.V1.Domain;
 using ContactDetailsApi.V1.Factories;
-using ContactDetailsApi.V1.Gateways;
+using ContactDetailsApi.V1.Gateways.Interfaces;
 using ContactDetailsApi.V1.UseCase;
 using FluentAssertions;
 using Moq;
