+++ ContactDetailsApi.Tests/V2/Gateway/DynamoDbGatewayTests.cs
@@ -20,16 +20,16 @@ namespace ContactDetailsApi.Tests.V2.Gateway
     public class DynamoDbGatewayTests : IDisposable
     {
         private readonly Fixture _fixture = new Fixture();
-        private readonly Mock<ILogger<DynamoDbGateway>> _logger;
+        private readonly Mock<ILogger<ContactDetailsDynamoDbGateway>> _logger;
         private readonly IDynamoDbFixture _dbFixture;
-        private readonly DynamoDbGateway _classUnderTest;
+        private readonly ContactDetailsDynamoDbGateway _classUnderTest;
         private readonly List<Action> _cleanup = new List<Action>();
 
         public DynamoDbGatewayTests(MockWebApplicationFactory<Startup> appFactory)
         {
-            _logger = new Mock<ILogger<DynamoDbGateway>>();
+            _logger = new Mock<ILogger<ContactDetailsDynamoDbGateway>>();
             _dbFixture = appFactory.DynamoDbFixture;
-            _classUnderTest = new DynamoDbGateway(_dbFixture.DynamoDbContext, _logger.Object);
+            _classUnderTest = new ContactDetailsDynamoDbGateway(_dbFixture.DynamoDbContext, _logger.Object);
         }
 
         public void Dispose()
