+++ ContactDetailsApi.Tests/V2/UseCase/CreateContactUseCaseTests.cs
@@ -4,7 +4,8 @@ using ContactDetailsApi.V1.Domain.Sns;
 using ContactDetailsApi.V2.Boundary.Response;
 using ContactDetailsApi.V2.Domain;
 using ContactDetailsApi.V2.Factories;
-using ContactDetailsApi.V2.Gateways;
+using ContactDetailsApi.V2.Factories.Interfaces;
+using ContactDetailsApi.V2.Gateways.Interfaces;
 using ContactDetailsApi.V2.Infrastructure;
 using ContactDetailsApi.V2.UseCase;
 using FluentAssertions;
