+++ ContactDetailsApi.Tests/V2/UseCase/GetContactDetailsByTargetIdUseCaseTest.cs
@@ -3,7 +3,7 @@ using ContactDetailsApi.V1.Boundary.Request;
 using ContactDetailsApi.V2.Boundary.Response;
 using ContactDetailsApi.V2.Domain;
 using ContactDetailsApi.V2.Factories;
-using ContactDetailsApi.V2.Gateways;
+using ContactDetailsApi.V2.Gateways.Interfaces;
 using ContactDetailsApi.V2.UseCase;
 using FluentAssertions;
 using Moq;
