+++ ContactDetailsApi/V1/UseCase/CreateContactUseCase.cs
@@ -1,7 +1,8 @@
 using ContactDetailsApi.V1.Boundary.Request;
 using ContactDetailsApi.V1.Boundary.Response;
 using ContactDetailsApi.V1.Factories;
-using ContactDetailsApi.V1.Gateways;
+using ContactDetailsApi.V1.Factories.Interfaces;
+using ContactDetailsApi.V1.Gateways.Interfaces;
 using ContactDetailsApi.V1.Infrastructure;
 using ContactDetailsApi.V1.UseCase.Interfaces;
 using Hackney.Core.JWT;
